Current Progress at getting it to build and run on macOS
The changes to meshacces.cpp are a kludge to get around an error for "taking the address of a temporary object"

The changes to intrule.cpp/hpp are iffy; need to be looked over and likely changed.

must be build with USE_NATIVE_ARCH=OFF

This probably broke things, nothing but building on apple silicon has been tested.

some code is by JSchoerbel, some by mhochsteger, and some by me.